+++ lovebird-api/src/main/kotlin/com/lovebird/api/dto/param/external/DiaryUploadPresignedUrlParam.kt
@@ -1,4 +1,4 @@
-package com.lovebird.api.dto.param
+package com.lovebird.api.dto.param.external
 
 data class DiaryUploadPresignedUrlParam(
 	val userId: Long,
