+++ lovebird-api/src/main/kotlin/com/lovebird/api/service/external/PresignedUrlService.kt
@@ -1,7 +1,7 @@
 package com.lovebird.api.service.external
 
-import com.lovebird.api.dto.param.DiaryUploadPresignedUrlParam
-import com.lovebird.api.dto.param.ProfileUploadPresignedUrlParam
+import com.lovebird.api.dto.param.external.DiaryUploadPresignedUrlParam
+import com.lovebird.api.dto.param.external.ProfileUploadPresignedUrlParam
 import com.lovebird.api.dto.response.PresignedUrlListResponse
 import com.lovebird.api.dto.response.PresignedUrlResponse
 import com.lovebird.common.utils.FilenameFormatter
